共通機能を基底クラスに切り出す
