+++ src/Code.ts
@@ -3,7 +3,7 @@ import ToranoanaChecklistGenerator from "./ToranoanaChecklistGenerator";
 
 function main() {
     const checklistGenerator = new ToranoanaChecklistGenerator();
-    checklistGenerator.saveProductsFromTranoanaPromotionMail();
+    checklistGenerator.saveProductsFromPromotionMail();
     const melonbooksChecklistGenerator = new MelonbooksChecklistGenerator();
     melonbooksChecklistGenerator.saveProductsFromPromotionMail();
 }
