+++ src/IShinkanDraft.ts
@@ -1,0 +1,5 @@
+export interface IShinkanDraft {
+    title: string;
+    circleName: string;
+    url: string;
+}
